requirements.txt
numpy
pandas

matplotlib
seaborn

python-box
pyYAML
joblib

mlflow
dvc
dagshub

jupyter
ipykernel